# docs/swagger-inline.yaml
openapi: 3.0.0
paths:
...
  /departments:
    get:
      operationId: DepartmentsController_findAll
      summary: 获取部门列表
      parameters: []
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    description: 部门列表
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: number
                          description: 部门ID
                          example: 1
                        name:
                          type: string
                          description: 部门名称
                          example: 技术部
                        description:
                          type: string
                          description: 部门描述
                          example: 负责技术研发工作
                        parent_id:
                          type: number
                          description: 父部门ID
                          example: 1
                        sort_order:
                          type: number
                          description: 排序
                          example: 0
                        status:
                          type: number
                          description: 状态
                          example: 1
                        employeeCount:
                          type: number
                          description: 员工数量
                          example: 10
                        created_at:
                          format: date-time
                          type: string
                          description: 创建时间
                          example: '2023-01-01T00:00:00Z'
                        updated_at:
                          format: date-time
                          type: string
                          description: 更新时间
                          example: '2023-01-01T00:00:00Z'
                      required:
                        - id
                        - name
                        - sort_order
                        - status
                        - created_at
                        - updated_at
                required:
                  - data
        '401':
          description: 未授权
      tags:
        - 部门管理
      security:
        - bearer: []
  /departments/{id}:
    get:
      operationId: DepartmentsController_findOne
      summary: 获取部门详情
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    description: 部门ID
                    example: 1
                  name:
                    type: string
                    description: 部门名称
                    example: 技术部
                  description:
                    type: string
                    description: 部门描述
                    example: 负责技术研发工作
                  parent_id:
                    type: number
                    description: 父部门ID
                    example: 1
                  sort_order:
                    type: number
                    description: 排序
                    example: 0
                  status:
                    type: number
                    description: 状态
                    example: 1
                  employeeCount:
                    type: number
                    description: 员工数量
                    example: 10
                  created_at:
                    format: date-time
                    type: string
                    description: 创建时间
                    example: '2023-01-01T00:00:00Z'
                  updated_at:
                    format: date-time
                    type: string
                    description: 更新时间
                    example: '2023-01-01T00:00:00Z'
                required:
                  - id
                  - name
                  - sort_order
                  - status
                  - created_at
                  - updated_at
        '401':
          description: 未授权
        '404':
          description: 部门不存在
      tags:
        - 部门管理
      security:
        - bearer: []
  /roles:
    get:
      operationId: RolesController_findAll
      summary: 获取角色列表
      parameters: []
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    description: 角色列表
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: number
                          description: 角色ID
                          example: 1
                        code:
                          type: string
                          description: 角色代码
                          example: admin
                        name:
                          type: string
                          description: 角色名称
                          example: 系统管理员
                        description:
                          type: string
                          description: 角色描述
                          example: 系统管理员角色
                        permissions:
                          description: 权限列表
                          example:
                            - user:read
                            - user:write
                          type: array
                          items:
                            type: string
                        status:
                          type: number
                          description: 状态
                          example: 1
                        created_at:
                          format: date-time
                          type: string
                          description: 创建时间
                          example: '2023-01-01T00:00:00Z'
                        updated_at:
                          format: date-time
                          type: string
                          description: 更新时间
                          example: '2023-01-01T00:00:00Z'
                      required:
                        - id
                        - code
                        - name
                        - status
                        - created_at
                        - updated_at
                required:
                  - data
        '401':
          description: 未授权
      tags:
        - 角色管理
      security:
        - bearer: []
  /roles/{id}:
    get:
      operationId: RolesController_findOne
      summary: 获取角色详情
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '200':
          description: 获取成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: number
                    description: 角色ID
                    example: 1
                  code:
                    type: string
                    description: 角色代码
                    example: admin
                  name:
                    type: string
                    description: 角色名称
                    example: 系统管理员
                  description:
                    type: string
                    description: 角色描述
                    example: 系统管理员角色
                  permissions:
                    description: 权限列表
                    example:
                      - user:read
                      - user:write
                    type: array
                    items:
                      type: string
                  status:
                    type: number
                    description: 状态
                    example: 1
                  created_at:
                    format: date-time
                    type: string
                    description: 创建时间
                    example: '2023-01-01T00:00:00Z'
                  updated_at:
                    format: date-time
                    type: string
                    description: 更新时间
                    example: '2023-01-01T00:00:00Z'
                required:
                  - id
                  - code
                  - name
                  - status
                  - created_at
                  - updated_at
        '401':
          description: 未授权
        '404':
          description: 角色不存在
      tags:
        - 角色管理
      security:
        - bearer: []
...
info:
  title: OKR绩效考核系统API
  description: OKR绩效考核系统后端API文档
  version: '1.0'
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: bearer
      bearerFormat: JWT
      type: http
...